Rewroked config class so that options settings the config files are handled correctly.  This involved a rewrite of have the config code.  Also changed "rem-all-" to "clear-" and added support for other prefixes.
Reworked filter code:

  Reworked error handling.

  Eleminted seperate ... encoder ... deconder .. filter

  Removed notifier when adding filters
    filter modules are added as needed not when the "add-filter"
    option is set

  Removed short option hack in new_fmode.cpp as it was completly
  unnecesaary.  The exiting code in aspell.cpp works just fine.  I do
  not intend to allow users to specify custom short options for modes.

  Removed overly complicated Directoyr/PathBrowser

  Removed option "option-path".  option files are
  expected to be in the same location the library is found

+++ common/string.cpp
@@ -38,7 +38,7 @@ namespace acommon {
     int res;
     va_list ap;
   loop: {
-      size_t avail = storage_end_ - end_;
+      int avail = storage_end_ - end_;
       if (avail > 1024) return -1;
       va_copy(ap,ap0);
       res = vsnprintf(end_, avail, format, ap);
